requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==4.0.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.2
django-paypal==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
martor==1.6.26
oauthlib==3.2.2
paypalrestsdk==1.13.1
paystack==1.5.0
paystackapi==2.1.0
Pillow==9.5.0
platformdirs==3.8.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
webencodings==0.5.1
whitenoise==6.5.0
